# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VERA GATE Payment Gateway API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [MENUNGGU, SUKSES, GAGAL, KEDALUWARSA]
        - name: merchantId
          in: query
          schema:
            type: ["integer", "null"]
        - name: method
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /transactions/generate:
    post:
      operationId: generateTransaction
      tags: [transactions]
      summary: Generate QRIS for a new topup transaction (20-min cooldown if MENUNGGU exists)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction created with QRIS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionWithQr"
        "409":
          description: Cooldown active — existing MENUNGGU transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CooldownError"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/sukses:
    get:
      operationId: listSuksesTransactions
      tags: [transactions]
      summary: List only SUKSES (successful) transactions
      parameters:
        - name: merchantId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of successful transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /transactions/callback:
    post:
      operationId: transactionCallback
      tags: [transactions]
      summary: Flypay webhook callback — marks transaction SUKSES if valid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallbackInput"
      responses:
        "200":
          description: Callback processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{id}/status:
    patch:
      operationId: updateTransactionStatus
      tags: [transactions]
      summary: Manually update transaction status (Status Lock prevents changing SUKSES)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "409":
          description: Status lock — cannot change SUKSES transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{ref}/check:
    post:
      operationId: checkTransactionStatus
      tags: [transactions]
      summary: Manually check transaction status with Flypay and update if needed
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    User:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, operator, merchant]
        merchantId:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, operator, merchant]
        merchantId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        password:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, operator, merchant]
        merchantId:
          type: ["integer", "null"]
    Transaction:
      type: object
      required: [id, ref, amount, status, method, createdAt]
      properties:
        id:
          type: integer
        ref:
          type: string
        amount:
          type: integer
        status:
          type: string
          enum: [MENUNGGU, SUKSES, GAGAL, KEDALUWARSA]
        method:
          type: string
        qrCode:
          type: ["string", "null"]
        customerId:
          type: ["string", "null"]
        merchantId:
          type: ["integer", "null"]
        merchantName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TransactionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TransactionInput:
      type: object
      required: [amount, customerId]
      properties:
        amount:
          type: integer
          minimum: 1000
        customerId:
          type: string
        merchantId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    TransactionWithQr:
      type: object
      required: [transaction, qrCode, ref]
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        qrCode:
          type: string
        ref:
          type: string
    CooldownError:
      type: object
      required: [error, cooldownMinutes, existingRef, existingCreatedAt]
      properties:
        error:
          type: string
        cooldownMinutes:
          type: integer
        remainingMinutes:
          type: integer
        existingRef:
          type: string
        existingCreatedAt:
          type: string
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [MENUNGGU, SUKSES, GAGAL, KEDALUWARSA]
        notes:
          type: ["string", "null"]
    CallbackInput:
      type: object
      properties:
        order:
          type: ["string", "null"]
        orderId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        transaction_id:
          type: ["string", "null"]
        uid:
          type: ["string", "null"]
...