# lib/api-zod/src/generated/types/callbackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERA GATE Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface CallbackInput {
  /** @nullable */
  order?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  transaction_id?: string | null;
  /** @nullable */
  uid?: string | null;
}