# lib/api-zod/src/generated/types/transactionWithQr.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERA GATE Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface TransactionWithQr {
  transaction: Transaction;
  qrCode: string;
  ref: string;
}